kafka-producer/requirements.txt
kafka-python
aiohttp
